+++ lib/cocoapods-core/podfile/target_definition.rb
@@ -468,6 +468,7 @@ module Pod
       # @return [void]
       #
       def store_pod(name, *requirements)
+        return if parse_subspecs(name, requirements) # This parse method must be called first
         parse_inhibit_warnings(name, requirements)
         parse_configuration_whitelist(name, requirements)
 
@@ -741,6 +742,29 @@ module Pod
           end
         end
         requirements.pop if options.empty?
+      end
+
+      # Removes :subspecs form the requirements list, and stores the pods
+      # with the given subspecs as dependencies.
+      #
+      # @param  [String] name
+      #
+      # @param  [Array] requirements
+      #         If :subspecs is the only key in the hash, the hash
+      #         should be destroyed because it confuses Gem::Dependency.
+      #
+      # @return [Boolean] Whether new subspecs were added
+      #
+      def parse_subspecs(name, requirements)
+        options = requirements.last
+        return false unless options.is_a?(Hash)
+
+        subspecs = options.delete(:subspecs)
+        subspecs.each do |ss|
+          store_pod("#{name}/#{ss}", *requirements.dup)
+        end if subspecs
+        requirements.pop if options.empty?
+        !subspecs.nil?
       end
 
       #-----------------------------------------------------------------------#