+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Master
 
+##### Enhancements
+
+* Add support for test specifications in the podspec DSL.  
+  [Kyle Fuller](https://github.com/kylef)
+  [#142](https://github.com/CocoaPods/Core/pull/142)
+
 ##### Bug Fixes
 
 * Inhibit warnings for pods that only have the `inhibit_warnings` option enabled