+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ##### Enhancements
 
 * Allow specifying multiple subspec pod dependencies inline in the Podfile,
-  via `pod 'Pod', subspecs: ['Subspec1', 'Subspec2']`.  
+  via `pod 'Pod', :subspecs => ['Subspec1', 'Subspec2']`.  
   [Samuel Giddins](https://github.com/segiddins)
   [#221](https://github.com/CocoaPods/Core/issues/221)