+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CocoaPods Core Changelog
 
+## 0.36.4
+
+This version only introduces changes in the CocoaPods gem.
+
+
 ## 0.36.3
 
 This version only introduces changes in the CocoaPods gem.